Massive Commit (read Description)
Added better Damage all enemies action (for mod compatibility)

Added RelicPreviewButton to applicable events

Rewrote Snecko Tail
Added Stabinomicon and added patches to make it appear in Cursed Tome event and small cahnce after fighting it.
Fixed Sticky Explosives
Added Tome of Questing (infinite crossover)
Made Future Diary compulsion card more obvious.
Made Hummingbird Hear uncommon
Changed Infernal Blood (removed power heal, inc attack heal)
Changed Kaom's Heart's price
Removed RelicAboveCreature effect from marisa Skillbook
Removed Envenom Powertip from Silent Skillbook
Reworked Snecko Tail (Befuddled)
Allowed SCoin to trigger whenever you have charges and fixed pulsing
Reworked a few relic images

Made all powers Cloneable
Added Beffuddled and Grievous Wounds Powers because of relics.

+++ src/main/java/aspiration/relics/EnhancedActuators.java
@@ -37,9 +37,9 @@ public static final String ID = "aspiration:EnhancedActuators";
 	public void onOrbSlotChange(int amount, boolean isReduce) {
 		int tmp = 0;
 		if (isReduce) {
-			tmp = (int) Math.floor((AbstractDungeon.player.maxOrbs - amount) / ORB_SLOT_THRESHOLD);
+			tmp = (int) Math.floor(((float) AbstractDungeon.player.maxOrbs - amount) / ORB_SLOT_THRESHOLD);
 		} else {
-			tmp = (int) Math.floor((AbstractDungeon.player.maxOrbs + amount) / ORB_SLOT_THRESHOLD);
+			tmp = (int) Math.floor(((float) AbstractDungeon.player.maxOrbs + amount) / ORB_SLOT_THRESHOLD);
 		}
 
 		if (current_strength_given > tmp) {
